# arm64: optimize base64 decode block by replacing ORR + SHL + SHR with SLI + SHR