This is an attempt to reduce the instruction count from 11 to 9 as seen in the below Compiler Explorer link
https://godbolt.org/z/Krze65Kcs

Before:
```
./build/benchmarks/base64/benchmark_base64 -f simdutf -d base64data/email/*
# current system detected as arm64.
# loading files: ...............
# volume: 1976492 bytes
# max length: 1006349 bytes
# number of inputs: 15
# decode
# the base64 data contains spaces, so we cannot use straigth libbase64::base64_decode directly
simdutf::arm64                           :   7.53 GB/s  2.30 %
simdutf::arm64 (accept garbage)          :   7.58 GB/s  1.90 %
```

After:
```
./build2/benchmarks/base64/benchmark_base64 -f simdutf -d base64data/email/*
# current system detected as arm64.
# loading files: ...............
# volume: 1976492 bytes
# max length: 1006349 bytes
# number of inputs: 15
# decode
# the base64 data contains spaces, so we cannot use straigth libbase64::base64_decode directly
simdutf::arm64                           :   7.75 GB/s  2.33 %
simdutf::arm64 (accept garbage)          :   7.73 GB/s  2.35 %
```