@lemire 2025-04-01
I am merging because it is obviously good.